Fix flickering sky if time is too high (Closes #200)
This was caused because the max int size is 2,147,483,647, which is what Minecraft: Bedrock Edition uses in the time packet. In Minecraft: Java Edition, a long is used which has a max length of 9,223,372,036,854,775,807, thus causing the sky to bug out.

This commit uses the modulus operator with the max time value per-day of 24,000 to retrieve the remainder.