Correct the block state if PlayerActionAckPacket sends it
Originally we did not translate if this packet failed. With this commit, a lot (if not all) of the ghost block issues with mobile devices should be solved if the player is running a new enough version that supports this packet. If the packet returns not successful, we reset the block back to its original type.

This commit also includes cleanup in BlockChangeTranslator as well as proper Netherite tool support for calculating block breaking.