Add world border translation
As Bedrock does not have world border support, this PR translates what the Java server sends us for a world border into particles and a message, while also preventing the Bedrock client from moving outside of the world border.
Thanks to both Luke and OFunny who started on this work.

Co-authored-by: Luke <32024335+lukeeey@users.noreply.github.com>
Co-authored-by: ofunny <play@ofunny.world>