Don't use Locale class in MessageTranslator
The Minecraft Java client doesn't appear to use any specific locale function in its translating. Just passing a string saves computational time in parsing a locale that will be the same result every time.
Should this be invalid behavior, we should instead cache the Locale class for a player so it doesn't have to be repeatedly calculated.