close threads at server shutdown
To exit normally all non-daemon threads of the java process
are supposed to have stopped. There are several cases of
EventLoopGroup-created thread pools that are not shut down
as a part of the server shutdown.

These threads are spawned on the first Bedrock player login. The
identified cases are:
- The listener threads in GeyserServerInitializer
- The player threads in GeyserServerInitializer
- The EXECUTOR_SERVICE in the SkinProvider

This patch depends on the PR
https://github.com/GeyserMC/MCProtocolLib/pull/748
submitted to MCProtocolLib to allow the TcpCLientSession to
shut down on process exit.

This is a very quickly thrown together proof of concept to verify that
the execution thread pools left running were what blocked a clean exit
with Geyser active.

Nullify EventLoopGroup variables after shutdown to allow them to be
restarted.