Split GeyserBootstrap disable and shutdown logic
- Don't disable extensions at reload
- Fire GeyserPreReloadEvent and GeyserPostReloadEvent instead of the pre and post init events so extensions can do custom reloading

Unfortunately, we can't disable and enable extensions  without consequences - e.g. it is not possible to add or remove commands during runtime. Hence, it makes more sense to let extensions running during reloads, and provide them with information about the reload so extensions can implement custom reload handling.

In the future, some Geyser lifecycle events could be re-fired; for example the custom item or block registries. This however is outside the scope of this PR.

Further changes:
- split geyser.shutdown() into disable and shutdown respectively to better separate between reloading and stopping
- undo extension hot-loading
- don't re-create the GeyserImpl instance on fabric with reloads
- Listen to spigot's serverloadevent to determine if we need to end our shutdown process and start running again
- Add GeyserPreReloadEvent and GeyserPostReloadEvent to indicate reload status to extensions

TODO:
- testing all bootstraps